refactor(desktop): split app shell layout actions
