+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -8,6 +8,7 @@ const sourcePaths = [
   'main.tsx',
   'app.tsx',
   'app-shell-turn-actions.ts',
+  'app-shell-layout-actions.ts',
   'app-shell.tsx',
   'app-shell-plan-actions.ts',
   'app-shell-skill-actions.ts',
